Move ApplePay to ApplePayParams
