+++ KsApi/mutations/inputs/UpdateBackingInput.swift
@@ -2,34 +2,15 @@ import Foundation
 
 public struct UpdateBackingInput: GraphMutationInput, Encodable {
   let amount: String?
-  let applePay: ApplePay?
+  let applePay: ApplePayParams?
   let id: String
   let locationId: String?
   let paymentSourceId: String?
   let rewardId: String?
 
-  public struct ApplePay: Encodable {
-    let paymentInstrumentName: String
-    let paymentNetwork: String
-    let transactionIdentifier: String
-    let token: String
-
-    public init(
-      paymentInstrumentName: String,
-      paymentNetwork: String,
-      transactionIdentifier: String,
-      token: String
-    ) {
-      self.paymentInstrumentName = paymentInstrumentName
-      self.paymentNetwork = paymentNetwork
-      self.transactionIdentifier = transactionIdentifier
-      self.token = token
-    }
-  }
-
   public init(
     amount: String?,
-    applePay: ApplePay?,
+    applePay: ApplePayParams?,
     id: String,
     locationId: String?,
     paymentSourceId: String?,
@@ -41,9 +22,5 @@ public struct UpdateBackingInput: GraphMutationInput, Encodable {
     self.locationId = locationId
     self.paymentSourceId = paymentSourceId
     self.rewardId = rewardId
-  }
-
-  public func toInputDictionary() -> [String: Any] {
-    return self.dictionaryRepresentation ?? [:]
   }
 }
